+++ packages/playlist/src/index.ts
@@ -1,3 +1,0 @@
-export { createPlaylistClient } from './playlist';
-
-export type { components } from './playlistAPI';